updated shapes, began loading bar, modded text

+++ planner/ui/components/impl/shapes/Shape.java
@@ -6,11 +6,11 @@ public interface Shape {
 
     public double getArea();
     public double getPerimeter();
-    public boolean contains(int x, int y);
+    public boolean contains(double x, double y);
     public boolean contains(Point p);
     public boolean contains(Polygon s);
     public boolean intersects(Polygon s);
-    public void translate(int x, int y);
+    public void translate(double x, double y);
     public void translate(Point p);
     public void scale(double factor);
     public void skew(double xFactor, double yFactor);
